// src/types/categories.ts
export type ProductCategoryType = {
    id: string;
    category_name: string;
    slug: string;
    description: string
    metaTitle: string
    metaDescription: string
}



export type BlogCategoryType = {
    id: string;
    category_name: string;
    slug: string;
    description: string
    metaTitle: string
    metaDescription: string
}